\name{joint.Cox-package}
\alias{joint.Cox-package}
\alias{joint.Cox}
\docType{package}
\title{
Penalized Likelihood Estimation and Prediction under the Joint Cox Models Between Tumour Progression and
 Death for Meta-Analysis
}
\description{
Perform regression analyses under the joint Cox proportional hazards model between tumour progression and death for meta-analysis.
 The method is applicable for meta-analysis combining several studies or for cluster survival data.
}
\details{
\tabular{ll}{
Package: \tab joint.Cox\cr
Type: \tab Package\cr
Version: \tab 2.2\cr
Date: \tab 2015-10-03\cr
License: \tab GPL-2\cr
}
}
\author{
Takeshi Emura 
Maintainer: Takeshi Emura <takeshiemura@gmail.com>
}
\references{
Emura T*, Nakatochi M, Murotani K, Rondeau V (2015), A joint frailty-copula model between tumour progression and death for meta-analysis,
 Statistical Methods in Medical Research, doi: 10.1177/0962280215604510

}
\keyword{ package }

