\name{joint.Cox-package}
\alias{joint.Cox-package}
\alias{joint.Cox}
\docType{package}
\title{
Penalized Likelihood Estimation under the Joint Cox Models Between TTP 
 and OS for Meta-Analysis
}
\description{
Perform regression analyses under the joint Cox proportional hazards model between TTP and OS for meta-analysis.
 The method is applicable for meta-analysis combining several studies or for cluster survival data.
}
\details{
\tabular{ll}{
Package: \tab joint.Cox\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-05-24\cr
License: \tab GPL-2\cr
}
}
\author{
Takeshi Emura 
Maintainer: Takeshi Emura <emura@stat.ncu.edu.tw>
}
\references{
Emura T*, Nakatochi M, Murotani K, Rondeau V (2015), A joint frailty-copula model between tumour progression and death for meta-analysis,
under revision, Statistical Methods in Medical Research.
}
\keyword{ package }

