\name{joint.Cox-package}
\alias{joint.Cox-package}
\alias{joint.Cox}
\docType{package}
\title{
Joint Frailty-Copula Models for Tumour Progression and Death in Meta-Analysis
}
\description{
Perform likelihood estimation and dynamic prediction under joint frailty-copula models for tumour progression and death in meta-analysis.
 A penalized likelihood method is employed for estimating model parameters, where the baseline hazard functions are modeled by smoothing splines.
 The methods are applicable for meta-analytic data combining several studies.
 The methods can analyze data having information on both terminal event time (e.g., time-to-death) and non-terminal event time (e.g., time-to-tumour progression).
 See Emura et al. (2017), Emura et al. (2018), and Emura et al. (2020-) for more details.
 More details on these methods can also be found in a book of Emura et al. (2019).
 Survival data from ovarian cancer patients are also available.
}
\details{
\tabular{ll}{
Package: \tab joint.Cox\cr
Type: \tab Package\cr
Version: \tab 3.7\cr
Date: \tab 2019-11-20\cr
License: \tab GPL-2\cr
}
}
\author{
Takeshi Emura 
Maintainer: Takeshi Emura <takeshiemura@gmail.com>
}
\references{
Emura T, Nakatochi M, Murotani K, Rondeau V (2017), A joint frailty-copula model between tumour progression and death for meta-analysis,
 Stat Methods Med Res 26(6): 2649-66

Emura T, Nakatochi M, Matsui S, Michimae H, Rondeau V (2018), Personalized dynamic prediction of death according to tumour progression and high-dimensional
 genetic factors: meta-analysis with a joint model, Stat Methods Med Res 27(9):2842-58

Emura T, Matsui S, Rondeau V (2019), Survival Analysis with Correlated Endpoints, Joint Frailty-Copula Models, JSS Research Series in Statistics, Springer

Emura T, Shih JH, Ha ID, Wilke RA (2020-), Comparison of the Marginal Hazard Model and the Sub-distribution Hazard Model for Competing Risks under an Assumed Copula,
 Stat Methods Med Res, in press
}
\keyword{ package }

