\name{joint.Cox-package}
\alias{joint.Cox-package}
\alias{joint.Cox}
\docType{package}
\title{
Penalized Likelihood Estimation and Dynamic Prediction under the Joint Frailty-Copula Models Between Tumour Progression and
 Death for Meta-Analysis
}
\description{
Perform the Cox regression and dynamic prediction methods under the joint frailty-copula model between tumour progression and death for meta-analysis.
 A penalized likelihood is employed for estimating model parameters, where the baseline hazard functions are approximated by smoothing splines.
 The methods are applicable for meta-analytic data combining several studies.
 The methods can analyze data having information on both terminal event time (e.g., time-to-death) and non-terminal event time (e.g., time-to-tumour progression).
 See Emura et al. (2015) and Emura et al. (2017) for details.
 Survival data from ovarian cancer patients are also available.
}
\details{
\tabular{ll}{
Package: \tab joint.Cox\cr
Type: \tab Package\cr
Version: \tab 2.12\cr
Date: \tab 2017-5-1\cr
License: \tab GPL-2\cr
}
}
\author{
Takeshi Emura 
Maintainer: Takeshi Emura <takeshiemura@gmail.com>
}
\references{
Emura T, Nakatochi M, Murotani K, Rondeau V (2015), A joint frailty-copula model between tumour progression and death for meta-analysis,
 Statistical Methods in Medical Research, doi: 10.1177/0962280215604510

Emura T, Nakatochi M, Matsui S, Michimae H, Rondeau V (2017), Personalized dynamic prediction of death according to tumour progression and high-dimensional
 genetic factors: meta-analysis with a joint model Statistical Methods in Medical Research, doi:10.1177/0962280216688032
}
\keyword{ package }

