\name{joint.Cox-package}
\alias{joint.Cox-package}
\alias{joint.Cox}
\docType{package}
\title{
Penalized Likelihood Estimation and Dynamic Prediction under the Joint Frailty-Copula Models Between Tumour Progression and
 Death for Meta-Analysis
}
\description{
Perform the Cox regression and dynamic prediction under the joint frailty-copula model between tumour progression and death for meta-analysis.
 The method is applicable for meta-analytic data combining several studies. The data should have information on both terminal event time (time-to-death) and non-terminal event time (time-to-tumour progression).
}
\details{
\tabular{ll}{
Package: \tab joint.Cox\cr
Type: \tab Package\cr
Version: \tab 2.8\cr
Date: \tab 2016-9-30\cr
License: \tab GPL-2\cr
}
}
\author{
Takeshi Emura 
Maintainer: Takeshi Emura <takeshiemura@gmail.com>
}
\references{
Emura T*, Nakatochi M, Murotani K, Rondeau V (2015), A joint frailty-copula model between tumour progression and death for meta-analysis,
 Statistical Methods in Medical Research, doi: 10.1177/0962280215604510

Emura T*, Nakatochi M, Matsui S, Michimae H, Rondeau V (2016), Personalized dynamic prediction of death according to tumour progression and high-dimensional
 genetic factors: meta-analysis with a joint model, in revision, Statistical Methods in Medical Research.
}
\keyword{ package }

